requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.40.0
supabase==2.9.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic==2.9.2
